Add in MachineState to the Control trait, HTTP Endpoint (#118)
This is a first step as we start to chip away on #68 slash #49 and others.

This will add a standard part of our "Control" trait to check on the state of a printer -- is the printer running? idle? paused mid-print? Broken?

I implemented this for Moonraker (since I was able to test it), and stubbed in Unknown elsewhere. USB is doable here -- and I think we can do Bambu, but that may wait until I get into the office to play with it myself.